doip_ops: handle the checkalive while doing and UDS req
In the very tiny temporal window where a UDS request was sent, but before an UDS
acknownledge was received (not the response, the fist DoIP ack), if a
AliveCheckRequest arrives, the UDS handling was broken.

Fix this by using in the send_uds() function the capability to handle also
sparse alive checks, and answer to them.

It is not clear so far this is fine with regard to the specification, as an
alive check should come in during a silence (ie. no UDS ongoing request), but a
timings cross will have this effect.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>